pyproject.toml
[project]
name = "bugfix-workflow"
version = "3.1.0"
description = "通用 AI Bug 修復工作流程引擎 — Skill-Based，支援 Copilot / Claude / OpenAI SDK"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [
    {name = "Bugfix Workflow Contributors"}
]

dependencies = [
    "pydantic>=2.12.5",        # config.py 使用 Pydantic 驗證 YAML 配置
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0",
    "requests>=2.32.5",
    "click>=8.1.0",            # CLI 工具
]

[project.optional-dependencies]
# 選擇其中一個 SDK 安裝即可
copilot = ["github-copilot-sdk>=0.1.20"]
claude  = ["anthropic>=0.40.0"]
openai  = ["openai-agents>=0.0.3"]

# 安裝全部（開發/測試用）
all = [
    "github-copilot-sdk>=0.1.20",
    "anthropic>=0.40.0",
    "openai-agents>=0.0.3",
]

dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
bugfix-workflow = "main:cli"
bfw = "main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=engine --cov-report=term-missing"

[tool.coverage.run]
source = ["engine"]
omit = ["tests/*", ".venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "@abstractmethod",
]
